+++ module/api/action/ship/ship-rotate-action.js
@@ -1,4 +1,5 @@
 import { getInitiatorToken } from "../../targeting.js";
+import { getTokenRotation } from "../../utils.js";
 
 /**
  * @param {PBActor} actor
@@ -7,7 +8,7 @@ import { getInitiatorToken } from "../../targeting.js";
 export const shipRotateAction = async (actor, angle) => {
   const token = getInitiatorToken(actor);
   if (token?.document) {
-    const currentRotation = token.data.rotation;
+    const currentRotation = getTokenRotation(token);
     await token.document.update({ rotation: currentRotation + angle });
   }
 };