+++ module/system/combat.js
@@ -1,4 +1,4 @@
-import { getSystemFlag, setSystemFlag } from "../api/utils.js";
+import { getCombatantInitiative, getSystemFlag, getTokenDisposition, setSystemFlag } from "../api/utils.js";
 
 /**
  * @extends Combat
@@ -16,7 +16,7 @@ export class PBCombat extends Combat {
     await this.updatePartyInitiative(rollTotal);
     const updates = this.turns.map((t) => ({
       _id: t.id,
-      initiative: t.data.initiative,
+      initiative: getCombatantInitiative(t),
     }));
     await game.combat.resetAll();
     await this.updateEmbeddedDocuments("Combatant", updates);
@@ -25,8 +25,8 @@ export class PBCombat extends Combat {
   /** @private */
   _sortCombatants(a, b) {
     const combat = a.combat; // game.combats is not initialized at first render
-    const isFriendlyA = a.token?.data.disposition === 1 ?? false;
-    const isFriendlyB = b.token?.data.disposition === 1 ?? false;
+    const isFriendlyA = (a.token && getTokenDisposition(a.token)) === 1 ?? false;
+    const isFriendlyB = (b.token && getTokenDisposition(b.token)) === 1 ?? false;
     const isVehicleA = a.actor?.isAnyVehicle ?? false;
     const isVehicleB = b.actor?.isAnyVehicle ?? false;